+++ db/migrate/20170302172001_add_last_backup_at_to_users.rb
@@ -1,0 +1,6 @@
+class AddLastBackupAtToUsers < ActiveRecord::Migration[5.0]
+  def change
+    add_column :users, :last_backup_at, :datetime
+    add_column :users, :last_backup_entry_id, :integer, null: false, default: 0
+  end
+end